Rename cocotb.log to cocotb.logging
