+++ docs/source/newsfragments/3673.change.rst
@@ -1,0 +1,1 @@
+The module ``cocotb.log`` was renamed to :module:`cocotb.logging` to prevent clashing with :attr:`cocotb.log`.
