+++ docs/source/newsfragments/3673.removal.rst
@@ -1,1 +1,1 @@
-Removed :class:`~cocotb.log.SimLog`, use :meth:`logging.getLogger` instead.
+Removed ``cocotb.logging.SimLog``, use :meth:`logging.getLogger` instead.
