+++ tests/test_cases/test_cocotb/test_logging.py
@@ -9,7 +9,7 @@ import logging
 import os
 
 import cocotb
-import cocotb.log
+import cocotb.logging
 import pytest
 
 
@@ -43,7 +43,7 @@ async def test_logging_with_args(dut):
 async def test_logging_default_config(dut):
     # The cocotb.log module is shadowed by an instance of
     # cocotb.log.SimBaseLog()
-    from cocotb.log import default_config as log_default_config
+    from cocotb.logging import default_config as log_default_config
 
     cocotb_log = logging.getLogger("cocotb")
