+++ tests/test_cases/test_custom_entry/custom_entry.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from cocotb.log import _filter_from_c, _log_from_c  # noqa: F401
+from cocotb.logging import _filter_from_c, _log_from_c  # noqa: F401
 
 
 def entry_func(argv: List[str]) -> None:
